refactor(android): modernize activities to compose screens and remove legacy shims
