--- app/src/main/java/com/vinnovateit/latch/domain/model/WifiStatsManager.kt
@@ -1,6 +1,0 @@
-package com.vinnovateit.latch.domain.model
-
-typealias DataUsage = com.vinnovateit.latch.core.model.DataUsage
-typealias LiveDataPoint = com.vinnovateit.latch.core.model.LiveDataPoint
-typealias LiveConnectionStatus = com.vinnovateit.latch.core.model.LiveConnectionStatus
-typealias SessionSummary = com.vinnovateit.latch.core.model.SessionSummary
