+++ core/build.gradle.kts
@@ -76,4 +76,9 @@ android {
         sourceCompatibility = JavaVersion.VERSION_11
         targetCompatibility = JavaVersion.VERSION_11
     }
+
+    lint {
+        abortOnError = false
+        checkReleaseBuilds = false
+    }
 }
